Move private helper methods into signatureUtils so that they don't show up in the ZeroEx classes auto-complete list
